Emoji and color choice are being implemented to trackerCreationView
