+++ HabitsTracker/Modules/TrackersCreation/Subviews/EmojiCell.swift
@@ -1,0 +1,34 @@
+import UIKit
+
+final class EmojiCell: UICollectionViewCell {
+    private var emojiLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        addSubviews()
+        applyConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func addSubviews() {
+        contentView.addSubview(emojiLabel)
+    }
+    private func applyConstraints() {
+        NSLayoutConstraint.activate([
+            emojiLabel.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
+            emojiLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
+        ])
+    }
+    
+    func setEmoji(emoji: String) {
+        emojiLabel.text = emoji
+    }
+}
